# Support []int and []string in JSON encoder