--- a/field_test.go
+++ b/field_test.go
@@ -174,6 +174,13 @@ func TestMarshalerField(t *testing.T) {
 	assertCanBeReused(t, Marshaler("foo", fakeUser{"phil"}))
 }
 
+func TestIntsField(t *testing.T) {
+	assertFieldJSON(t, `"foo":[]`, Object("foo", []int{}))
+	assertFieldJSON(t, `"foo":[1]`, Object("foo", []int{1}))
+	assertFieldJSON(t, `"foo":[1,2,3]`, Object("foo", []int{1, 2, 3}))
+	assertCanBeReused(t, Object("foo", []int{1, 2, 3}))
+}
+
 func TestObjectField(t *testing.T) {
 	assertFieldJSON(t, `"foo":[5,6]`, Object("foo", []int{5, 6}))
 	assertCanBeReused(t, Object("foo", []int{5, 6}))