--- a/json_encoder.go
+++ b/json_encoder.go
@@ -95,7 +95,7 @@ func (enc *jsonEncoder) Free() {
 	jsonPool.Put(enc)
 }
 
-func (enc *jsonEncoder) wrapString(val string) {
+func (enc *jsonEncoder) addAndQuoteString(val string) {
 	enc.bytes = append(enc.bytes, '"')
 	enc.safeAddString(val)
 	enc.bytes = append(enc.bytes, '"')
@@ -105,7 +105,7 @@ func (enc *jsonEncoder) wrapString(val string) {
 // value are JSON-escaped.
 func (enc *jsonEncoder) AddString(key, val string) {
 	enc.addKey(key)
-	enc.wrapString(val)
+	enc.addAndQuoteString(val)
 }
 
 // AddBool adds a string key and a boolean value to the encoder's fields. The
@@ -172,15 +172,18 @@ func (enc *jsonEncoder) AddMarshaler(key string, obj LogMarshaler) error {
 	return err
 }
 
-func (enc *jsonEncoder) arrayBegin() {
+func (enc *jsonEncoder) addArrayBegin() {
 	enc.bytes = append(enc.bytes, '[')
 }
 
-func (enc *jsonEncoder) arrayEnd() {
+func (enc *jsonEncoder) addArrayEnd() {
 	enc.bytes = append(enc.bytes, ']')
 }
 
-func (enc *jsonEncoder) arraySep() {
+func (enc *jsonEncoder) addArraySep(i int) {
+	if i == 0 {
+		return
+	}
 	enc.bytes = append(enc.bytes, ',')
 }
 
@@ -188,34 +191,25 @@ func (enc *jsonEncoder) arraySep() {
 // is JSON-escaped.
 func (enc *jsonEncoder) AddInts(key string, vals []int) {
 	enc.addKey(key)
-	enc.arrayBegin()
-	l := len(vals)
-	if l > 0 {
-		enc.bytes = strconv.AppendInt(enc.bytes, int64(vals[0]), 10)
+	enc.addArrayBegin()
 
-		for i := 1; i < l; i++ {
-			enc.arraySep()
-			enc.bytes = strconv.AppendInt(enc.bytes, int64(vals[i]), 10)
-		}
+	for i := range vals {
+		enc.addArraySep(i)
+		enc.bytes = strconv.AppendInt(enc.bytes, int64(vals[i]), 10)
 	}
-	enc.arrayEnd()
+	enc.addArrayEnd()
 }
 
 // AddStrings adds a string key and string slice to the encoder's fields. The key
 // is JSON-escaped.
 func (enc *jsonEncoder) AddStrings(key string, vals []string) {
 	enc.addKey(key)
-	enc.arrayBegin()
-	l := len(vals)
-	if l > 0 {
-		enc.wrapString(vals[0])
-
-		for i := 1; i < l; i++ {
-			enc.arraySep()
-			enc.wrapString(vals[i])
-		}
+	enc.addArrayBegin()
+	for i := range vals {
+		enc.addArraySep(i)
+		enc.addAndQuoteString(vals[i])
 	}
-	enc.arrayEnd()
+	enc.addArrayEnd()
 }
 
 // AddObject uses reflection to add an arbitrary object to the logging context.