--- a/keyvalue.go
+++ b/keyvalue.go
@@ -29,6 +29,7 @@ type KeyValue interface {
 	AddBool(key string, value bool)
 	AddFloat64(key string, value float64)
 	AddInt(key string, value int)
+	AddInts(key string, value []int)
 	AddInt64(key string, value int64)
 	AddUint(key string, value uint)
 	AddUint64(key string, value uint64)
@@ -38,8 +39,5 @@ type KeyValue interface {
 	// allocation-heavy. Consider implementing the LogMarshaler interface instead.
 	AddObject(key string, value interface{}) error
 	AddString(key, value string)
-
-	// Slices
-	AddInts(key string, value []int)
 	AddStrings(key string, value []string)
 }