--- a/text_encoder.go
+++ b/text_encoder.go
@@ -116,7 +116,7 @@ func (enc *textEncoder) arrayEnd() {
 }
 
 func (enc *textEncoder) arraySep() {
-	enc.bytes = append(enc.bytes, ',')
+	enc.bytes = append(enc.bytes, ' ')
 }
 
 func (enc *textEncoder) AddInts(key string, vals []int) {
@@ -134,22 +134,16 @@ func (enc *textEncoder) AddInts(key string, vals []int) {
 	enc.arrayEnd()
 }
 
-func (enc *textEncoder) wrapString(val string) {
-	enc.bytes = append(enc.bytes, `"`...)
-	enc.bytes = append(enc.bytes, val...)
-	enc.bytes = append(enc.bytes, `"`...)
-}
-
 func (enc *textEncoder) AddStrings(key string, vals []string) {
 	enc.addKey(key)
 	enc.arrayBegin()
 	l := len(vals)
 	if l > 0 {
-		enc.wrapString(vals[0])
+		enc.bytes = append(enc.bytes, vals[0]...)
 
 		for i := 1; i < l; i++ {
 			enc.arraySep()
-			enc.wrapString(vals[i])
+			enc.bytes = append(enc.bytes, vals[i]...)
 		}
 	}
 	enc.arrayEnd()