--- a/text_encoder.go
+++ b/text_encoder.go
@@ -107,46 +107,39 @@ func (enc *textEncoder) AddMarshaler(key string, obj LogMarshaler) error {
 }
 
 // Helpers for Add Array fns below
-func (enc *textEncoder) arrayBegin() {
+func (enc *textEncoder) addArrayBegin() {
 	enc.bytes = append(enc.bytes, '[')
 }
 
-func (enc *textEncoder) arrayEnd() {
+func (enc *textEncoder) addArrayEnd() {
 	enc.bytes = append(enc.bytes, ']')
 }
 
-func (enc *textEncoder) arraySep() {
+func (enc *textEncoder) addArraySep(i int) {
+	if i == 0 {
+		return
+	}
 	enc.bytes = append(enc.bytes, ' ')
 }
 
 func (enc *textEncoder) AddInts(key string, vals []int) {
 	enc.addKey(key)
-	enc.arrayBegin()
-	l := len(vals)
-	if l > 0 {
-		enc.bytes = strconv.AppendInt(enc.bytes, int64(vals[0]), 10)
-
-		for i := 1; i < l; i++ {
-			enc.arraySep()
-			enc.bytes = strconv.AppendInt(enc.bytes, int64(vals[i]), 10)
-		}
+	enc.addArrayBegin()
+	for i := range vals {
+		enc.addArraySep(i)
+		enc.bytes = strconv.AppendInt(enc.bytes, int64(vals[i]), 10)
 	}
-	enc.arrayEnd()
+	enc.addArrayEnd()
 }
 
 func (enc *textEncoder) AddStrings(key string, vals []string) {
 	enc.addKey(key)
-	enc.arrayBegin()
-	l := len(vals)
-	if l > 0 {
-		enc.bytes = append(enc.bytes, vals[0]...)
-
-		for i := 1; i < l; i++ {
-			enc.arraySep()
-			enc.bytes = append(enc.bytes, vals[i]...)
-		}
+	enc.addArrayBegin()
+	for i := range vals {
+		enc.addArraySep(i)
+		enc.bytes = append(enc.bytes, vals[i]...)
 	}
-	enc.arrayEnd()
+	enc.addArrayEnd()
 }
 
 func (enc *textEncoder) AddObject(key string, obj interface{}) error {