I've implemented zap.Ints and zap.Strings and submitting this PR for discussion around API and output.

For API, see examples in logger_bench_test.go.

For output, the JSON and Text encoder outputs are equal to the current output when using zap.Object( []int / []string ).  See json_encoder_test.go and text_encoder_test.go.