@phungs 2016-12-13
Comparison of the json_array_pre branch (same as master) vs this branch using -benchtime 500ms, -cpu 1 and 20-35 iterations for each code branch, done before the rebase to json_array_pre.

It appears the new branch is generally faster, but I would take that with a grain of salt.  I would say only the Ints and Strings array tests are much faster, since they have a very wide delta (i.e >= 30%) and both bounds (±) are small/med (i.e. <= ~10%).

The other tests probably perform the same.  For tests where the delta and both bounds are small (i.e <= 5%) it's probably a wash/benchmark noise since the changes in this branch didn't affect those benchmarks.  And for a test like StringerField where the bounds are 12% and 3% and the delta is -9.27%, I would say it's inconclusive.

```
$ benchstat bm_json_array_pre_stat_cpu1_1 bm_json_array_obj_stat_cpu1_1
name                               old time/op    new time/op    delta
CheckedMessage_Chain                  238ns ± 5%     225ns ± 3%   -5.60%  (p=0.000 n=32+18)
CheckedMessage_Chain_sliceLoggers     241ns ± 5%     232ns ± 4%   -3.94%  (p=0.000 n=33+20)
JSONLogMarshalerFunc                  118ns ± 6%     117ns ± 6%     ~     (p=0.075 n=32+20)
ZapJSONInts                          1.15µs ± 4%    0.51µs ± 5%  -55.80%  (p=0.000 n=31+20)
ZapJSONStrings                       1.68µs ± 9%    0.56µs ± 3%  -66.49%  (p=0.000 n=30+20)
ZapJSON                               650ns ± 7%     637ns ± 2%   -1.97%  (p=0.026 n=34+19)
StandardJSON                         6.57µs ± 9%    5.91µs ± 2%  -10.02%  (p=0.000 n=33+19)
NullLogMarshalerFunc                 44.3ns ± 9%    42.5ns ± 4%   -4.12%  (p=0.000 n=32+20)
ZapNull                              39.5ns ± 7%    38.6ns ± 5%   -2.26%  (p=0.008 n=32+20)
NoContext                             545ns ±10%     536ns ± 4%     ~     (p=0.220 n=34+20)
BoolField                             652ns ± 6%     628ns ± 4%   -3.67%  (p=0.000 n=34+20)
Float64Field                          815ns ± 7%     768ns ± 2%   -5.74%  (p=0.000 n=33+15)
IntField                              669ns ± 6%     642ns ± 2%   -4.01%  (p=0.000 n=35+20)
Int64Field                            699ns ± 9%     644ns ± 1%   -7.93%  (p=0.000 n=34+17)
StringField                           667ns ± 7%     635ns ± 3%   -4.75%  (p=0.000 n=35+19)
StringerField                         748ns ±12%     679ns ± 3%   -9.27%  (p=0.000 n=35+20)
TimeField                             743ns ± 9%     675ns ± 2%   -9.26%  (p=0.000 n=35+17)
DurationField                         693ns ± 5%     664ns ± 2%   -4.10%  (p=0.000 n=35+19)
ErrorField                            682ns ± 6%     659ns ± 3%   -3.40%  (p=0.000 n=33+19)
StackField                           8.38µs ± 5%    8.12µs ± 2%   -3.08%  (p=0.000 n=35+19)
MarshalerField                        876ns ± 4%     851ns ± 3%   -2.87%  (p=0.000 n=35+20)
**
IntsField                            1.55µs ± 8%    0.96µs ± 7%  -38.26%  (p=0.000 n=35+20)
StringsField                         2.07µs ± 7%    0.98µs ± 3%  -52.62%  (p=0.000 n=35+19)
**
ObjectField                          2.51µs ± 4%    2.47µs ± 3%   -1.87%  (p=0.000 n=34+20)
AddCallerHook                        1.51µs ± 4%    1.45µs ± 2%   -3.88%  (p=0.000 n=33+19)
10Fields                             1.40µs ± 4%    1.39µs ± 2%     ~     (p=0.359 n=31+20)
100Fields                            6.72µs ± 7%    6.56µs ± 3%   -2.38%  (p=0.002 n=35+20)
Tee_Check                            1.27µs ±10%    1.23µs ± 6%   -2.83%  (p=0.021 n=35+20)

name                               old alloc/op   new alloc/op   delta
CheckedMessage_Chain                  42.0B ± 0%     42.0B ± 0%     ~     (all equal)
CheckedMessage_Chain_sliceLoggers     42.0B ± 0%     42.0B ± 0%     ~     (all equal)
JSONLogMarshalerFunc                  16.0B ± 0%     16.0B ± 0%     ~     (all equal)
ZapJSONInts                            296B ± 0%       80B ± 0%  -72.97%  (p=0.000 n=35+20)
ZapJSONStrings                         520B ± 0%      160B ± 0%  -69.23%  (p=0.000 n=35+20)
ZapJSON                               0.00B          0.00B          ~     (all equal)
StandardJSON                         1.92kB ± 0%    1.92kB ± 0%     ~     (all equal)
NullLogMarshalerFunc                  16.0B ± 0%     16.0B ± 0%     ~     (all equal)
ZapNull                               0.00B          0.00B          ~     (all equal)
NoContext                             0.00B          0.00B          ~     (all equal)
BoolField                             64.0B ± 0%     64.0B ± 0%     ~     (all equal)
Float64Field                          64.0B ± 0%     64.0B ± 0%     ~     (all equal)
IntField                              64.0B ± 0%     64.0B ± 0%     ~     (all equal)
Int64Field                            64.0B ± 0%     64.0B ± 0%     ~     (all equal)
StringField                           64.0B ± 0%     64.0B ± 0%     ~     (all equal)
StringerField                         68.0B ± 0%     68.0B ± 0%     ~     (all equal)
TimeField                             64.0B ± 0%     64.0B ± 0%     ~     (all equal)
DurationField                         64.0B ± 0%     64.0B ± 0%     ~     (all equal)
ErrorField                            64.0B ± 0%     64.0B ± 0%     ~     (all equal)
StackField                           1.09kB ± 0%    1.09kB ± 0%     ~     (all equal)
MarshalerField                        64.0B ± 0%     64.0B ± 0%     ~     (all equal)
**
IntsField                              360B ± 0%      176B ± 0%  -51.11%  (p=0.000 n=35+20)
StringsField                           584B ± 0%      256B ± 0%  -56.16%  (p=0.000 n=35+20)
**
ObjectField                            584B ± 0%      584B ± 0%     ~     (all equal)
AddCallerHook                         48.0B ± 0%     48.0B ± 0%     ~     (all equal)
10Fields                               640B ± 0%      640B ± 0%     ~     (all equal)
100Fields                            1.17kB ± 0%    1.17kB ± 0%     ~     (all equal)
Tee_Check                             64.0B ± 0%     64.0B ± 0%     ~     (all equal)

name                               old allocs/op  new allocs/op  delta
CheckedMessage_Chain                   0.00           0.00          ~     (all equal)
CheckedMessage_Chain_sliceLoggers      0.00           0.00          ~     (all equal)
JSONLogMarshalerFunc                   1.00 ± 0%      1.00 ± 0%     ~     (all equal)
ZapJSONInts                            4.00 ± 0%      1.00 ± 0%  -75.00%  (p=0.000 n=35+20)
ZapJSONStrings                         5.00 ± 0%      1.00 ± 0%  -80.00%  (p=0.000 n=35+20)
ZapJSON                                0.00           0.00          ~     (all equal)
StandardJSON                           38.0 ± 0%      38.0 ± 0%     ~     (all equal)
NullLogMarshalerFunc                   1.00 ± 0%      1.00 ± 0%     ~     (all equal)
ZapNull                                0.00           0.00          ~     (all equal)
NoContext                              0.00           0.00          ~     (all equal)
BoolField                              1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Float64Field                           1.00 ± 0%      1.00 ± 0%     ~     (all equal)
IntField                               1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Int64Field                             1.00 ± 0%      1.00 ± 0%     ~     (all equal)
StringField                            1.00 ± 0%      1.00 ± 0%     ~     (all equal)
StringerField                          2.00 ± 0%      2.00 ± 0%     ~     (all equal)
TimeField                              1.00 ± 0%      1.00 ± 0%     ~     (all equal)
DurationField                          1.00 ± 0%      1.00 ± 0%     ~     (all equal)
ErrorField                             1.00 ± 0%      1.00 ± 0%     ~     (all equal)
StackField                             2.00 ± 0%      2.00 ± 0%     ~     (all equal)
MarshalerField                         1.00 ± 0%      1.00 ± 0%     ~     (all equal)
**
IntsField                              5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=35+20)
StringsField                           6.00 ± 0%      3.00 ± 0%  -50.00%  (p=0.000 n=35+20)
**
ObjectField                            7.00 ± 0%      7.00 ± 0%     ~     (all equal)
AddCallerHook                          1.00 ± 0%      1.00 ± 0%     ~     (all equal)
10Fields                               1.00 ± 0%      1.00 ± 0%     ~     (all equal)
100Fields                              3.00 ± 0%      3.00 ± 0%     ~     (all equal)
Tee_Check                              1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```